backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
python-dotenv==1.0.1
openai==1.61.0
yt-dlp==2025.1.26
Werkzeug==3.0.4
requests==2.31.0
